# docs/swagger.yaml
definitions:
  models.Author:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        example: Mark Twains
        type: string
      status:
        example: enabled
        type: string
      updated_at:
        type: string
    required:
    - name
    - status
    type: object
  models.Book:
    properties:
      ISBN:
        example: "4564231"
        type: string
      authorID:
        example: uuid
        type: string
      categoryID:
        example: uuid
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      locationID:
        example: uuid
        type: string
      name:
        example: War and Peace
        type: string
      quantity:
        example: "10"
        type: string
      status:
        example: enabled
        type: string
      updated_at:
        type: string
    required:
    - ISBN
    - authorID
    - categoryID
    - locationID
    - name
    - quantity
    - status
    type: object
...
  models.CreateAuthorModul:
    properties:
      name:
        example: Mark Twains
        type: string
    required:
    - name
    type: object
  models.CreateBookModul:
    properties:
      ISBN:
        example: 4564231
        type: integer
      authorID:
        example: uuid
        type: string
      categoryID:
        example: uuid
        type: string
      locationID:
        example: uuid
        type: string
      name:
        example: War and Peace
        type: string
      quantity:
        example: 10
        type: integer
    required:
    - ISBN
    - authorID
    - categoryID
    - locationID
    - name
    - quantity
    type: object
...
  models.JSONErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.JSONResult:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  models.UpdateBookModul:
    properties:
      id:
        type: string
      locationID:
        example: uuid
        type: string
      name:
        example: War and Peace
        type: string
      quantity:
        example: 10
        type: integer
    required:
    - locationID
    - name
    - quantity
    type: object
...
info:
  contact: {}
paths:
  /v1/author:
    post:
      consumes:
      - application/json
      description: Creat a new author
      parameters:
      - description: Author body
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/models.CreateAuthorModul'
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/models.Author'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: Creat Author
      tags:
      - author
  /v1/author/:
    get:
      consumes:
      - application/json
      description: GetAuthorList
      parameters:
      - description: "0"
        in: query
        name: offset
        type: integer
      - description: "100"
        in: query
        name: limit
        type: integer
      - description: search exapmle
        in: query
        name: search
        type: string
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Author'
                  type: array
              type: object
      summary: List authors
      tags:
      - author
  /v1/author/{id}:
    delete:
      consumes:
      - application/json
      description: get element by id and delete this author
      parameters:
      - description: Author id
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/models.Author'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: Delete Author
      tags:
      - author
    get:
      consumes:
      - application/json
      description: get an author by id
      parameters:
      - description: Author id
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/models.Author'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: GetAuthorByID
      tags:
      - author
    put:
      consumes:
      - application/json
      description: get element by id and enabled this author
      parameters:
      - description: Author id
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/models.Author'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: Enabled Author
      tags:
      - author
  /v1/book:
    post:
      consumes:
      - application/json
      description: Creat a new book
      parameters:
      - description: Book body
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.CreateBookModul'
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/models.Book'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: Creat Book
      tags:
      - book
  /v1/book/:
    get:
      consumes:
      - application/json
      description: GetBookList
      parameters:
      - description: "0"
        in: query
        name: offset
        type: integer
      - description: "100"
        in: query
        name: limit
        type: integer
      - description: search exapmle
        in: query
        name: search
        type: string
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Book'
                  type: array
              type: object
      summary: List books
      tags:
      - book
    put:
      consumes:
      - application/json
      description: Update Book
      parameters:
      - description: Book body
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBookModul'
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Book'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: Update Book
      tags:
      - book
  /v1/book/{id}:
    delete:
      consumes:
      - application/json
      description: get element by id and delete this book
      parameters:
      - description: Book id
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/models.Book'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: Delete Book
      tags:
      - book
    get:
      consumes:
      - application/json
      description: get an book by id
      parameters:
      - description: Book id
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/models.Book'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: GetBookByID
      tags:
      - book
    put:
      consumes:
      - application/json
      description: get element by id and delete this book
      parameters:
      - description: Book id
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/models.Book'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: Enabled Book
      tags:
      - book
...
swagger: "2.0"